+++ ga4gh/drs/config/global_state.py
@@ -1,0 +1,12 @@
+class GlobalState(object):
+
+    def __init__(self):
+        self.props = {}
+    
+    def set_prop(self, key, value):
+        self.props[key] = value
+    
+    def get_prop(self, key):
+        return self.props[key]
+
+GLOBALSTATE = GlobalState()