+++ ga4gh/drs/util/method_types/gs.py
@@ -19,16 +19,15 @@ class GS(MethodType):
         download_submethods (list): multiple methods to attempt byte download
     """
     
-    def __init__(self, json, drs_obj, cli_kwargs):
+    def __init__(self, json, drs_obj):
         """Instantiates a GS object
 
         Arguments:
             json (dict): parsed AccessMethod JSON, used to set other attributes
             drs_obj (DRSObject): reference to parent DRSObject object
-            cli_kwargs (dict): command-line arguments
         """
 
-        super(GS, self).__init__(json, drs_obj, cli_kwargs)
+        super(GS, self).__init__(json, drs_obj)
         self.download_submethods = [
             self.__download_by_https,
             self.__download_by_gsutil