+++ ga4gh/drs/util/method_types/http.py
@@ -15,16 +15,15 @@ class HTTP(MethodType):
         download_submethods (list): multiple methods to attempt byte download
     """
 
-    def __init__(self, json, drs_obj, cli_kwargs):
+    def __init__(self, json, drs_obj):
         """Instantiates an HTTP object
 
         Arguments:
             json (dict): parsed AccessMethod JSON, used to set other attributes
             drs_obj (DRSObject): reference to parent DRSObject object
-            cli_kwargs (dict): command-line arguments, options
         """
 
-        super(HTTP, self).__init__(json, drs_obj, cli_kwargs)
+        super(HTTP, self).__init__(json, drs_obj)
         self.download_submethods = [
             self.__download_by_http
         ]