+++ unittests/test_access_method.py
@@ -1,11 +1,14 @@
 import json
 from unittests.cli_kwargs import *
+from ga4gh.drs.config.global_state import GLOBALSTATE 
 from ga4gh.drs.definitions.object import DRSObject
 from ga4gh.drs.definitions.access_method import AccessMethod
+from ga4gh.drs.util.logger import Logger
 
-cli_kwargs = CLI_KWARGS_0
+GLOBALSTATE.set_prop("cli", CLI_KWARGS_0)
+GLOBALSTATE.set_prop("logger", Logger("drs.client.unittests"))
 drs_obj_json = json.load(open("unittests/testdata/json/drs_object_0.json", "r"))
-drs_obj = DRSObject(drs_obj_json, cli_kwargs)
+drs_obj = DRSObject(drs_obj_json)
 
 def test_access_method():