+++ unittests/test_data_accessor.py
@@ -14,13 +14,13 @@ from unittests.cli_kwargs import CLI_KWARGS_0
 
 drs_inner_bundle_json = open("unittests/testdata/json/drs_object_3.json", "r").read()
 drs_inner_bundle_obj = json.loads(drs_inner_bundle_json)
-drs_inner_bundle = DRSObject(drs_inner_bundle_obj, CLI_KWARGS_0)
+drs_inner_bundle = DRSObject(drs_inner_bundle_obj)
 
 def test_data_accessor():
 
     dt = DownloadTree(drs_inner_bundle)
     dt.recurse_find_leaves(drs_inner_bundle)
-    data_accessors = dt.get_data_accessors_for_leaves(CLI_KWARGS_0)
+    data_accessors = dt.get_data_accessors_for_leaves()
     checksum_statuses = [
         cs.ChecksumStatus.FAILED,
         cs.ChecksumStatus.PASSED,