+++ unittests/test_download_manager.py
@@ -13,7 +13,7 @@ from unittests.cli_kwargs import CLI_KWARGS_0
 
 drs_inner_bundle_json = open("unittests/testdata/json/drs_object_3.json", "r").read()
 drs_inner_bundle_obj = json.loads(drs_inner_bundle_json)
-drs_inner_bundle = DRSObject(drs_inner_bundle_obj, CLI_KWARGS_0)
+drs_inner_bundle = DRSObject(drs_inner_bundle_obj)
 
 exp_download_statuses = [
     ds.DownloadStatus.COMPLETED,
@@ -26,10 +26,10 @@ def test_download_manager():
 
     dt = DownloadTree(drs_inner_bundle)
     dt.recurse_find_leaves(drs_inner_bundle)
-    data_accessors = dt.get_data_accessors_for_leaves(CLI_KWARGS_0)
+    data_accessors = dt.get_data_accessors_for_leaves()
     assert len(data_accessors) == 4
 
-    dm = DownloadManager(data_accessors, CLI_KWARGS_0)
+    dm = DownloadManager(data_accessors)
     assert len(dm.data_accessors) == 4
     assert dm.max_workers == 1
     assert dm.report_file == "unittests/outdata/drs_download_report.txt"